+++ src/firebird/base/config.py
@@ -55,7 +55,10 @@ import platform
 from pathlib import Path
 from uuid import UUID
 from decimal import Decimal, DecimalException
-from configparser import ConfigParser, DEFAULTSECT
+from configparser import (ConfigParser, DEFAULTSECT, ExtendedInterpolation,
+                          MAX_INTERPOLATION_DEPTH, InterpolationDepthError,
+                          InterpolationSyntaxError, NoSectionError, NoOptionError,
+                          InterpolationMissingOptionError)
 from inspect import signature, Signature, Parameter
 from enum import Enum, Flag
 import os
@@ -139,6 +142,74 @@ def _power_of_two(value):
     if value < 1:
         return False
     return value == 2 ** (value.bit_length() - 1)
+
+class EnvExtendedInterpolation(ExtendedInterpolation):
+    """.. versionadded:: 1.8.0
+
+    Modified version of `configparser.ExtendedInterpolation` class that adds special
+    handling for "env" section that returns value of specified environment variable,
+    or empty string if such variable is not defined.
+
+    Example::
+
+       ${env:path} is reference to PATH environment variable.
+    """
+    def _interpolate_some(self, parser, option, accum, rest, section, map,
+                          depth):
+        rawval = parser.get(section, option, raw=True, fallback=rest)
+        if depth > MAX_INTERPOLATION_DEPTH:
+            raise InterpolationDepthError(option, section, rawval)
+        while rest:
+            p = rest.find("$")
+            if p < 0:
+                accum.append(rest)
+                return
+            if p > 0:
+                accum.append(rest[:p])
+                rest = rest[p:]
+            # p is no longer used
+            c = rest[1:2]
+            if c == "$":
+                accum.append("$")
+                rest = rest[2:]
+            elif c == "{":
+                m = self._KEYCRE.match(rest)
+                if m is None:
+                    raise InterpolationSyntaxError(option, section,
+                        "bad interpolation variable reference %r" % rest)
+                path = m.group(1).split(':')
+                rest = rest[m.end():]
+                sect = section
+                opt = option
+                try:
+                    if len(path) == 1:
+                        opt = parser.optionxform(path[0])
+                        v = map[opt]
+                    elif len(path) == 2:
+                        sect = path[0]
+                        opt = parser.optionxform(path[1])
+                        if sect == 'env':
+                            v = os.getenv(opt.upper(), '')
+                        else:
+                            v = parser.get(sect, opt, raw=True)
+                    else:
+                        raise InterpolationSyntaxError(
+                            option, section,
+                            "More than one ':' found: %r" % (rest,))
+                except (KeyError, NoSectionError, NoOptionError):
+                    raise InterpolationMissingOptionError(
+                        option, section, rawval, ":".join(path)) from None
+                if "$" in v:
+                    self._interpolate_some(parser, opt, accum, v, sect,
+                                           dict(parser.items(sect, raw=True)),
+                                           depth + 1)
+                else:
+                    accum.append(v)
+            else:
+                raise InterpolationSyntaxError(
+                    option, section,
+                    "'$' must be followed by '$' or '{', "
+                    "found: %r" % (rest,))
 
 class DirectoryScheme:
     """Class that provide paths to typically used application directories.