+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.8.0] - 2024-05-03
+
+### Added
+
+- New `EnvExtendedInterpolation` class in `cnfig` module that extends `configparser.ExtendedInterpolation`
+  with special handling for "env" section that returns value of specified environment
+  variable, or empty string if such variable is not defined.
+
 ## [1.7.2] - 2024-02-20
 
 ### Fixed